Set forward port/method on the virtual service
This simplifies configuration for real servers. A real server trying to
join the service only needs to add its IP.

The forwarding port and method are configured centrally on the service,
and applied to all attached real servers.

This is also more consistent with health checks being on the service.

It reduces flexibility of the real servers, but we believe this is a
worthwhile tradeoff. To support possibly adding real server ports back
in, the database key is left with the port embedded.