Add timeouts for netlink IPVS calls (#23)
We observed a netlink call hang indefinitely (despite libnetwork setting
a socket timeout). This caused the merlin reconciler loop to hang and
stop working. To guard against this rare occurrence, we wrap all netlink
calls in a goroutine.

In addition, merlin will now panic if the netlink call fails, as it's
unexpected to ever happen normally and may be unrecoverable.